updated easter egg hunting
