+++ .gitignore
@@ -1,1 +1,1 @@
-all_items.csv
+/results
